generic lru
